# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Loc-Nav",
        "version": "1.0.0"
    },
    "paths": {
        "/api/ships/location": {
            "get": {
                "tags": [
                    "Ships"
                ],
                "summary": "Get all ships locations",
                "operationId": "dcc7a92b6cde864b918c3b0ee0deac8b",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "imo": {
                                                "type": "integer",
                                                "example": 9943267
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "EVER ACME"
                                            },
                                            "flag": {
                                                "type": "string",
                                                "example": "Singapore"
                                            },
                                            "latitude": {
                                                "type": "string",
                                                "example": "-34"
                                            },
                                            "longitude": {
                                                "type": "string",
                                                "example": "26"
                                            },
                                            "external_system": {
                                                "type": "string",
                                                "example": "VESSEL_FINDER"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An unexpected error occurred. Please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ships/location/{imo}/{externalSystemId}": {
            "get": {
                "tags": [
                    "Ships"
                ],
                "summary": "Search ship location by IMO and external system",
                "operationId": "5e690073bfec1c68268b0fb4bb1f3b1e",
                "parameters": [
                    {
                        "name": "imo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 9839131
                        }
                    },
                    {
                        "name": "externalSystemId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "imo": {
                                            "type": "integer",
                                            "example": 9839131
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "CC CHAMPS ELYSEES"
                                        },
                                        "flag": {
                                            "type": "string",
                                            "example": "France"
                                        },
                                        "latitude": {
                                            "type": "string",
                                            "example": "-34"
                                        },
                                        "longitude": {
                                            "type": "string",
                                            "example": "18"
                                        },
                                        "external_system": {
                                            "type": "string",
                                            "example": "VESSEL_FINDER"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid external service",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid external service!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "IMO not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "IMO not found!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An unexpected error occurred. Please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Ships",
            "description": "Ships"
        }
    ]
}